- draft openspec api for batch ensembling - refactor REST controller - refactor ID usage of the model types

+++ api/turing/models/router_version_test.go
@@ -9,8 +9,8 @@ import (
 
 func TestRouterVersionSetters(t *testing.T) {
 	rv := RouterVersion{}
-	rv.SetEnricherID(uint(1))
-	rv.SetEnsemblerID(uint(2))
+	rv.SetEnricherID(1)
+	rv.SetEnsemblerID(2)
 	// Validate
 	assert.Equal(t, sql.NullInt32{Int32: int32(1), Valid: true}, rv.EnricherID)
 	assert.Equal(t, sql.NullInt32{Int32: int32(2), Valid: true}, rv.EnsemblerID)
